requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking
#
ankipandas==0.3.14
    # via -r requirements.in
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
colorlog==6.7.0
    # via ankipandas
duolingo-api @ git+https://github.com/GiulianoCTRL/Duolingo.git
    # via -r requirements.in
idna==3.4
    # via requests
numpy==1.24.3
    # via
    #   ankipandas
    #   pandas
pandas==2.0.1
    # via
    #   -r requirements.in
    #   ankipandas
python-dateutil==2.8.2
    # via pandas
pytz==2023.3
    # via pandas
randomfiletree==1.2.0
    # via ankipandas
requests==2.30.0
    # via
    #   -r requirements.in
    #   duolingo-api
six==1.16.0
    # via python-dateutil
tzdata==2023.3
    # via pandas
urllib3==2.0.2
    # via requests
